requirements.txt
fastapi
uvicorn
pydantic
jinja2
requests
openai
python-dotenv
httpx 